# Improve entity details audits query