--- a/lib/model/query/audits.js
+++ b/lib/model/query/audits.js
@@ -132,21 +132,9 @@ const getBySubmissionId = (submissionId, options) => ({ all }) => _getBySubmissi
   }));
 
 
-const _getByEntityId = (fields, options, entityId) => sql`
+const _getEntityDefsWithSubs = (fields, entityId) => sql`
 SELECT ${fields} FROM entity_defs
-
   LEFT JOIN entity_def_sources on entity_def_sources.id = entity_defs."sourceId"
-  INNER JOIN audits ON ((audits.details->>'entityDefId')::INTEGER = entity_defs.id OR (audits.details->>'sourceId')::INTEGER = entity_def_sources.id)
-
-  LEFT JOIN actors ON actors.id=audits."actorId"
-
-  LEFT JOIN audits triggering_event ON entity_def_sources."auditId" = triggering_event.id
-  LEFT JOIN actors triggering_event_actor ON triggering_event_actor.id = triggering_event."actorId"
-
-  -- if triggering event has a submissionId defined, look up creation event for that submission
-  -- it has info about the submission and creator we want to show even if the submission is deleted
-  LEFT JOIN audits submission_create_event ON (triggering_event.details->'submissionId')::INTEGER = (submission_create_event.details->'submissionId')::INTEGER AND submission_create_event.action = 'submission.create'
-  LEFT JOIN actors submission_create_event_actor ON submission_create_event_actor.id = submission_create_event."actorId"
 
   -- if source submissionDefId is defined:
   LEFT JOIN (
@@ -169,60 +157,90 @@ SELECT ${fields} FROM entity_defs
   -- if some other kind of target object defined, add subquery here
   -- ...
   WHERE entity_defs."entityId" = ${entityId}
+  ORDER BY entity_defs.id DESC;
+`;
+
+const _getByEntityId = (fields, options, entityId) => sql`
+SELECT ${fields} FROM entity_defs
+
+  LEFT JOIN entity_def_sources on entity_def_sources.id = entity_defs."sourceId"
+  INNER JOIN audits ON ((audits.details->>'entityDefId')::INTEGER = entity_defs.id OR (audits.details->>'sourceId')::INTEGER = entity_def_sources.id)
+
+  LEFT JOIN actors ON actors.id=audits."actorId"
+
+  LEFT JOIN audits triggering_event ON entity_def_sources."auditId" = triggering_event.id
+  LEFT JOIN actors triggering_event_actor ON triggering_event_actor.id = triggering_event."actorId"
+
+  -- if triggering event has a submissionId defined, look up creation event for that submission
+  -- it has info about the submission and creator we want to show even if the submission is deleted
+  LEFT JOIN audits submission_create_event ON (triggering_event.details->'submissionId')::INTEGER = (submission_create_event.details->'submissionId')::INTEGER AND submission_create_event.action = 'submission.create'
+  LEFT JOIN actors submission_create_event_actor ON submission_create_event_actor.id = submission_create_event."actorId"
+
+  -- if some other kind of target object defined, add subquery here
+  -- ...
+  WHERE entity_defs."entityId" = ${entityId}
   ORDER BY audits."loggedAt" DESC, audits.id DESC
   ${page(options)}`;
 
 const getByEntityId = (entityId, options) => ({ all }) => {
 
   const _unjoiner = unjoiner(
-    Audit, Actor, Entity.Def.Source,
+    Audit, Actor, Entity.Def, Entity.Def.Source,
     Option.of(Audit.alias('triggering_event', 'triggeringEvent')), Option.of(Actor.alias('triggering_event_actor', 'triggeringEventActor')),
     Option.of(Audit.alias('submission_create_event', 'submissionCreateEvent')), Option.of(Actor.alias('submission_create_event_actor', 'submissionCreateEventActor')),
+  );
+
+  const _defUnjoiner = unjoiner(
+    Entity.Def, Entity.Def.Source,
     Option.of(Submission), Option.of(Submission.Def.alias('current_submission_def', 'currentVersion')),
     Option.of(Actor.alias('current_submission_actor', 'currentSubmissionActor')),
     Option.of(Actor.alias('submission_actor', 'submissionActor')),
     Option.of(Form)
   );
 
-  return all(_getByEntityId(_unjoiner.fields, options, entityId))
-    .then(map(_unjoiner))
-    .then(map(audit => {
+  return Promise.all([
+    all(_getByEntityId(_unjoiner.fields, options, entityId)).then(map(_unjoiner)),
+    all(_getEntityDefsWithSubs(_defUnjoiner.fields, entityId)).then(map(_defUnjoiner))
+  ])
+    .then(([audits, defs]) => {
+      const entityDefDict = Object.fromEntries(defs.map(def => [def.id, def]));
 
-      const entitySourceDetails = audit.aux.source.forApi();
+      return audits.map(audit => {
+        const entitySourceDetails = audit.aux.source.forApi();
 
-      const sourceEvent = audit.aux.triggeringEvent
-        .map(a => a.withAux('actor', audit.aux.triggeringEventActor.orNull()))
-        .map(a => a.forApi());
+        const sourceEvent = audit.aux.triggeringEvent
+          .map(a => a.withAux('actor', audit.aux.triggeringEventActor.orNull()))
+          .map(a => a.forApi());
 
-      // If the entity event is based on a submission, get submission details from the submission create event,
-      // which should always exist, even if the entity has been deleted.
-      const submission = audit.aux.submissionCreateEvent.map((createEvent) => {
-        const baseSubmission = {
-          instanceId: createEvent.details.instanceId,
-          createdAt: createEvent.loggedAt,
-          submitter: audit.aux.submissionCreateEventActor.get().forApi()
-        };
+        // If the entity event is based on a submission, get submission details from the submission create event,
+        // which should always exist, even if the entity has been deleted.
+        const baseSubmission = audit.aux.submissionCreateEvent.map((createEvent) =>
+          ({
+            instanceId: createEvent.details.instanceId,
+            createdAt: createEvent.loggedAt,
+            submitter: audit.aux.submissionCreateEventActor.get().forApi()
+          }))
+          .orElse(undefined);
 
-        // If the submission has not been deleted, return full submission details, not just what
-        // was in the submission create event details.
-        const fullSubmission = audit.aux.submission
-          .map(s => s.withAux('submitter', audit.aux.submissionActor.orNull()))
-          .map(s => s.withAux('currentVersion', audit.aux.currentVersion.map(v => v.withAux('submitter', audit.aux.currentSubmissionActor.orNull()))))
+        // Looking up in a different dict for the full submission details if they exist
+        const subOption = entityDefDict[audit.aux.def.id];
+        const fullSubmission = subOption.aux.submission
+          .map(s => s.withAux('submitter', subOption.aux.submissionActor.orNull()))
+          .map(s => s.withAux('currentVersion', subOption.aux.currentVersion.map(v => v.withAux('submitter', subOption.aux.currentSubmissionActor.orNull()))))
           .map(s => s.forApi())
-          .map(s => mergeLeft(s, { xmlFormId: audit.aux.form.map(f => f.xmlFormId).orNull() }));
-        return mergeLeft(baseSubmission, fullSubmission.orElse(undefined));
-      })
-        .orElse(undefined);
+          .map(s => mergeLeft(s, { xmlFormId: subOption.aux.form.map(f => f.xmlFormId).orNull() }));
 
-      // Note: The source added to each audit event represents the source of the
-      // corresponding entity _version_, rather than the source of the event.
-      const details = mergeLeft(audit.details, sourceEvent
-        .map(event => ({ source: { event, submission } }))
-        .orElse({ source: entitySourceDetails }));
+        const submission = mergeLeft(baseSubmission, fullSubmission.orElse(undefined));
 
+        // Note: The source added to each audit event represents the source of the
+        // corresponding entity _version_, rather than the source of the event.
+        const details = mergeLeft(audit.details, sourceEvent
+          .map(event => ({ source: { event, submission } }))
+          .orElse({ source: entitySourceDetails }));
 
-      return new Audit({ ...audit, details }, { actor: audit.aux.actor });
-    }));
+        return new Audit({ ...audit, details }, { actor: audit.aux.actor });
+      });
+    });
 };