Hopefully addresses https://github.com/getodk/central/issues/618

I ended up pulling out the crazy part of this query that gets all the submission details that are vaguely relevant to an entity. It is now in another query that gets run separately and then if there is submission data to use, it is combined in code.

I'm curious to try this on test! I've been `EXPLAIN ANALYZE`-ing the queries. Hopefully an entity with a bulk source loads _super_ fast, single milliseconds.

### Entity 648 - created from a submission
https://test.getodk.cloud/#/projects/496/entity-lists/participants/entities/b5725bd9-fb5f-41e0-ba30-0eb8e6905e07

Original query (master)
* Planning Time: 10.761 ms
* Execution Time: 1357.950 ms
(I've seen it 800ms to >2 seconds)

New queries
Part 1: get all audit details and source events
* Planning Time: 2.093 ms
* Execution Time: 438.128 ms
(Seen between 300ms to 900ms)

Part 2: get full submission details
* Planning Time: 27.339 ms
* Execution Time: 2.709 ms
(this part is much faster w/o being joined with audits)

### Entity 118019 - created from bulk upload
https://test.getodk.cloud/#/projects/364/entity-lists/trees/entities/1d8cc1ba-fce5-44c9-acc2-e1fbafbaf81e

Original query
* Planning Time: 3.402 ms
* Execution Time: 515.838 ms
(I've seen it much slower though, 400-800ms, maybe longer)

New queries
Part 1
* Planning Time: 9.669 ms
* Execution Time: 5.648 ms

Part 2
* Planning Time: 8.170 ms
* Execution Time: 2.586 ms

<!-- 
Thank you for contributing to ODK Central!

Before sending this PR, please read
https://github.com/getodk/central-backend/blob/master/CONTRIBUTING.md
-->

#### What has been done to verify that this works as intended?

#### Why is this the best possible solution? Were any other approaches considered?

#### How does this change affect users? Describe intentional changes to behavior and behavior that could have accidentally been affected by code changes. In other words, what are the regression risks?

#### Does this change require updates to the API documentation? If so, please update docs/api.yaml as part of this PR.

#### Before submitting this PR, please make sure you have:

- [ ] run `make test` and confirmed all checks still pass OR confirm CircleCI build passes
- [ ] verified that any code from external sources are properly credited in comments or that everything is internally sourced