@ktuite 2024-04-18
I just tried rewriting the query to conceptually start with `audits` rather than `entity_defs`, but this appears to be slower, especially for entities that don't have a submission as a source.

```
SELECT ${fields} FROM audits
  LEFT JOIN actors ON actors.id=audits."actorId"

  LEFT JOIN entity_defs ON (audits.details->'entityDefId')::INTEGER = entity_defs.id
  LEFT JOIN entity_def_sources on entity_def_sources.id = entity_defs."sourceId"

  LEFT JOIN audits triggering_event ON entity_def_sources."auditId" = triggering_event.id
  LEFT JOIN actors triggering_event_actor ON triggering_event_actor.id = triggering_event."actorId"

  -- if triggering event has a submissionId defined, look up creation event for that submission
  -- it has info about the submission and creator we want to show even if the submission is deleted
  LEFT JOIN audits submission_create_event ON (triggering_event.details->'submissionId')::INTEGER = (submission_create_event.details->'submissionId')::INTEGER AND submission_create_event.action = 'submission.create'
  LEFT JOIN actors submission_create_event_actor ON submission_create_event_actor.id = submission_create_event."actorId"

  WHERE entity_defs."entityId" = ${entityId}
    AND (audits.details->>'entityDefId')::INTEGER = entity_defs.id OR (audits.details->>'sourceId')::INTEGER = entity_def_sources.id

  ORDER BY audits."loggedAt" DESC, audits.id DESC
```